+++ engine/src/main/java/dev/gamekit/ui/widgets/IntrinsicSize.java
@@ -1,20 +1,28 @@
 package dev.gamekit.ui.widgets;
 
 import dev.gamekit.ui.Constraints;
+import dev.gamekit.ui.enums.Axis;
 
 import java.util.Objects;
 
 /**
  * A {@link SingleChildParent} which limits the computed size of its single
- * child to the child's intrinsic size
+ * child to the child's intrinsic size in the specified {@link Axis}
  */
 public class IntrinsicSize extends SingleChildParent {
-  protected IntrinsicSize(Widget child) {
+  protected Axis axis;
+
+  protected IntrinsicSize(Axis axis, Widget child) {
     super(child);
+    this.axis = axis;
   }
 
   public static IntrinsicSize create(Widget child) {
-    return new IntrinsicSize(child);
+    return new IntrinsicSize(Axis.BOTH, child);
+  }
+
+  public static IntrinsicSize create(Axis axis, Widget child) {
+    return new IntrinsicSize(axis, child);
   }
 
   @Override
@@ -26,10 +34,20 @@ public class IntrinsicSize extends SingleChildParent {
       )
     );
 
-    intrinsicBounds.setSize(
-      child.intrinsicBounds.width,
-      child.intrinsicBounds.height
-    );
+    switch (axis) {
+      case VERTICAL -> intrinsicBounds.setSize(
+        child.intrinsicBounds.width,
+        child.intrinsicBounds.height
+      );
+      case HORIZONTAL -> intrinsicBounds.setSize(
+        child.intrinsicBounds.width,
+        child.computedBounds.height
+      );
+      case BOTH -> intrinsicBounds.setSize(
+        child.computedBounds.width,
+        child.intrinsicBounds.height
+      );
+    }
 
     computedBounds.setSize(
       constraints.constrainWidth(intrinsicBounds.width),
@@ -46,10 +64,16 @@ public class IntrinsicSize extends SingleChildParent {
 
   @Override
   protected boolean stateEquals(Widget widget) {
-    if (widget instanceof IntrinsicSize paddingWidget) {
-      return Objects.equals(child, paddingWidget.child);
+    if (widget instanceof IntrinsicSize intrinsicSizeWidget) {
+      return Objects.equals(child, intrinsicSizeWidget.child) &&
+        Objects.equals(axis, intrinsicSizeWidget.axis);
     }
 
     return false;
+  }
+
+  public IntrinsicSize withAxis(Axis axis) {
+    this.axis = axis;
+    return this;
   }
 }